Criado novo arquivo .c para o main

+++ vetor.h
@@ -12,6 +12,6 @@ v2d somaVetor(v2d v1, v2d v2);
 
 v2d subtracao(v2d v1, v2d v2);
 
-double modulo (v2d v);
+int modulo (v2d v);
 
 void mostrar(v2d n);
